Task #12231 change spetialType type in willow scheme
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@17083 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-model/src/main/java/ish/oncourse/model/auto/_Tag.java
@@ -108,11 +108,11 @@ public abstract class _Tag extends CayenneDataObject {
         return (String)readProperty("shortName");
     }
 
-    public void setSpecialType(Short specialType) {
+    public void setSpecialType(Integer specialType) {
         writeProperty("specialType", specialType);
     }
-    public Short getSpecialType() {
-        return (Short)readProperty("specialType");
+    public Integer getSpecialType() {
+        return (Integer)readProperty("specialType");
     }
 
     public void setWeighting(Integer weighting) {
